Squashed commit of the following:
commit 044007dac17d51c9e0b1b9e7b5e6329c30e6ff57
Author: Felipe Ferreira <feferreira.sf@gmail.com>
Date:   Wed Apr 1 18:35:41 2026 -0300

    Refactor ManageRefuelViewModel and Screen to improve state management by replacing toggle methods with direct setters for cold start and receipt photo

commit 1d5d94c0a66bc36e7884300219c1b6e55974dcca
Author: Felipe Ferreira <feferreira.sf@gmail.com>
Date:   Tue Mar 31 16:43:16 2026 -0300

    Refactor LoginViewModel and RegisterViewModel to simplify loading state management

commit 7e316e0175fc01bf0ab2355c168d30f350b0cb8c
Author: Felipe Ferreira <feferreira.sf@gmail.com>
Date:   Tue Mar 31 16:32:59 2026 -0300

    Refactor DashboardViewModel and VehicleCard to improve state management and simplify widget logic

commit b566e60fe398ad0c60c71de941c89536a2d6c276
Author: Felipe Ferreira <feferreira.sf@gmail.com>
Date:   Tue Mar 31 16:02:56 2026 -0300

    Refactor DashboardScreen and VehicleCard to simplify navigation and dismiss logic

commit b6092d40d46af87d791955bd0341fed4d35d4604
Author: Felipe Ferreira <feferreira.sf@gmail.com>
Date:   Tue Mar 31 15:57:04 2026 -0300

    Refactor ManageRefuelViewModel and Screen to enhance state management and improve editing logic

commit 096e1dceb4923580476b45d52cf73649826f3d0e
Author: Felipe Ferreira <feferreira.sf@gmail.com>
Date:   Tue Mar 31 15:44:19 2026 -0300

    Refactor ManageRefuelViewModel and Screen to implement delete confirmation dialog and improve state management

commit 819736d16f459c8da55fb802d56541aeea21855f
Author: Felipe Ferreira <feferreira.sf@gmail.com>
Date:   Tue Mar 31 15:24:30 2026 -0300

    Refactor ManageVehicleViewModel and Screen to streamline state management and improve editing logic

commit a2fa774ea7b83ec6511cd187367d799ea7627108
Author: Felipe Ferreira <feferreira.sf@gmail.com>
Date:   Tue Mar 31 14:53:51 2026 -0300

    Refactor ManageVehicleScreen and ViewModel to improve state management and loading logic

commit 5ec0db143da38620aa14b38528a09a3330598946
Author: Felipe Ferreira <feferreira.sf@gmail.com>
Date:   Tue Mar 31 14:16:15 2026 -0300

    Refactor VehicleDetailScreen and ViewModel to enhance state management and improve UI responsiveness

+++ lib/presentation/screens/refuel/manage_refuel_screen.dart
@@ -1,5 +1,3 @@
-import 'dart:io';
-
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:gasosa_app/core/app_strings.dart';
@@ -13,6 +11,7 @@ import 'package:gasosa_app/presentation/screens/refuel/viewmodel/manage_refuel_v
 import 'package:gasosa_app/presentation/widgets/gasosa_appbar.dart';
 import 'package:gasosa_app/presentation/widgets/gasosa_button.dart';
 import 'package:gasosa_app/presentation/widgets/gasosa_checkbox.dart';
+import 'package:gasosa_app/presentation/widgets/gasosa_confirm_dialog.dart';
 import 'package:gasosa_app/presentation/widgets/gasosa_date_picker_field.dart';
 import 'package:gasosa_app/presentation/widgets/gasosa_dropdown_field.dart';
 import 'package:gasosa_app/presentation/widgets/gasosa_form_field.dart';
@@ -40,23 +39,24 @@ class _ManageRefuelScreenState extends State<ManageRefuelScreen> {
   final _litersEC = TextEditingController();
   final _coldStartLitersEC = TextEditingController();
   final _coldStartValueEC = TextEditingController();
-  bool _didPopulate = false;
 
   @override
   void initState() {
     super.initState();
     _viewmodel = getIt<ManageRefuelViewModel>();
     _viewmodel.init(widget.refuelId, widget.vehicleId);
+    _viewmodel.loadCommand.state.addListener(_onLoadStateChanged);
   }
 
-  void _populateControllersIfNeeded(ManageRefuelState s) {
-    if (_didPopulate || !s.isEditing || s.initial == null) return;
+  void _onLoadStateChanged() {
+    if (_viewmodel.loadCommand.state.value is! UiData) return;
+    final s = _viewmodel.state.value;
     _mileageEC.text = NumericParser.formatInt(s.mileage);
     _totalValueEC.text = NumericParser.formatDouble(s.totalValue);
     _litersEC.text = NumericParser.formatDouble(s.liters);
     _coldStartLitersEC.text = s.coldStartLiters != null ? NumericParser.formatDouble(s.coldStartLiters!) : '';
     _coldStartValueEC.text = s.coldStartValue != null ? NumericParser.formatDouble(s.coldStartValue!) : '';
-    _didPopulate = true;
+    _viewmodel.loadCommand.state.removeListener(_onLoadStateChanged);
   }
 
   Future<void> _onSave() async {
@@ -73,12 +73,22 @@ class _ManageRefuelScreenState extends State<ManageRefuelScreen> {
   }
 
   Future<void> _onDelete() async {
+    final confirmed = await showGasosaConfirmDialog(
+      context,
+      title: RefuelStrings.deleteDialogTitle,
+      content: RefuelStrings.deleteDialogMessage,
+      confirmLabel: RefuelStrings.deleteDialogConfirmLabel,
+      danger: true,
+    );
+    if (!confirmed) return;
+
     final response = await _viewmodel.delete();
+    if (!mounted) return;
     response?.fold(
       (failure) => Messages.showError(context, failure.message),
       (_) {
         Messages.showSuccess(context, RefuelStrings.deleteSuccess);
-        if (mounted) context.pop(true);
+        context.pop(true);
       },
     );
   }
@@ -101,11 +111,11 @@ class _ManageRefuelScreenState extends State<ManageRefuelScreen> {
         ]),
         builder: (context, _) {
           final state = _viewmodel.state.value;
-          _populateControllersIfNeeded(state);
           final isLoading =
               _viewmodel.loadCommand.state.value is UiLoading ||
               _viewmodel.saveCommand.state.value is UiLoading ||
-              _viewmodel.deleteCommand.state.value is UiLoading;
+              _viewmodel.deleteCommand.state.value is UiLoading ||
+              _viewmodel.photoCommand.state.value is UiLoading;
 
           return SafeArea(
             child: SingleChildScrollView(
@@ -165,7 +175,7 @@ class _ManageRefuelScreenState extends State<ManageRefuelScreen> {
                         title: RefuelStrings.coldStartCheckboxLabel,
                         value: _viewmodel.hasColdStart,
                         onChanged: (value) {
-                          setState(() => _viewmodel.hasColdStart = value ?? false);
+                          setState(() => _viewmodel.setColdStart(value: value ?? false));
                           if (!(value ?? false)) {
                             _coldStartLitersEC.clear();
                             _coldStartValueEC.clear();
@@ -192,12 +202,12 @@ class _ManageRefuelScreenState extends State<ManageRefuelScreen> {
                     GasosaCheckbox(
                       title: RefuelStrings.receiptCheckboxLabel,
                       value: _viewmodel.hasReceiptPhoto,
-                      onChanged: (value) => setState(() => _viewmodel.hasReceiptPhoto = value ?? false),
+                      onChanged: (value) => setState(() => _viewmodel.toggleReceiptPhoto(value: value ?? false)),
                     ),
                     if (_viewmodel.shouldShowReceiptPhotoInput) ...[
                       GasosaPhotoPicker(
                         label: RefuelStrings.receiptPhotoLabel,
-                        image: state.receiptPath != null ? File(state.receiptPath!) : null,
+                        image: _viewmodel.currentReceiptPhoto,
                         onFileSelected: (file) async {
                           if (file == null) {
                             _viewmodel.onRemovePhoto();
@@ -220,7 +230,7 @@ class _ManageRefuelScreenState extends State<ManageRefuelScreen> {
                             onPressed: isLoading ? null : _onSave,
                           ),
                         ),
-                        if (state.isEditing) ...[
+                        if (_viewmodel.isEditing) ...[
                           Expanded(
                             child: GasosaButton(
                               label: RefuelStrings.deleteButton,
